QueryMetricsV2: Stream response
