+++ pkg/api/common.go
@@ -15,23 +15,6 @@ var (
 	}
 )
 
-type Response interface {
-	// WriteTo writes to the provided context.
-	WriteTo(ctx *models.ReqContext)
-	// Status gets the response's status.
-	Status() int
-	// Body gets the response's body.
-	Body() []byte
-}
-
-type NormalResponse struct {
-	status     int
-	body       []byte
-	header     http.Header
-	errMessage string
-	err        error
-}
-
 func Wrap(action interface{}) macaron.Handler {
 	return func(c *models.ReqContext) {
 		var res Response
@@ -46,44 +29,20 @@ func Wrap(action interface{}) macaron.Handler {
 	}
 }
 
-// Status gets the response's status.
-func (r *NormalResponse) Status() int {
-	return r.status
-}
-
-// Body gets the response's body.
-func (r *NormalResponse) Body() []byte {
-	return r.body
-}
-
-func (r *NormalResponse) WriteTo(ctx *models.ReqContext) {
-	if r.err != nil {
-		ctx.Logger.Error(r.errMessage, "error", r.err, "remote_addr", ctx.RemoteAddr())
-	}
-
-	header := ctx.Resp.Header()
-	for k, v := range r.header {
-		header[k] = v
-	}
-	ctx.Resp.WriteHeader(r.status)
-	if _, err := ctx.Resp.Write(r.body); err != nil {
-		ctx.Logger.Error("Error writing to response", "err", err)
-	}
-}
-
-func (r *NormalResponse) Header(key, value string) *NormalResponse {
-	r.header.Set(key, value)
-	return r
-}
-
-// Empty creates an empty response.
-func Empty(status int) *NormalResponse {
-	return Respond(status, nil)
-}
-
-// JSON create a JSON response
+// JSON creates a JSON response.
 func JSON(status int, body interface{}) *NormalResponse {
 	return Respond(status, body).Header("Content-Type", "application/json")
+}
+
+// jsonStreaming creates a streaming JSON response.
+func jsonStreaming(status int, body interface{}) streamingResponse {
+	header := make(http.Header)
+	header.Set("Content-Type", "application/json")
+	return streamingResponse{
+		status: status,
+		body:   body,
+		header: header,
+	}
 }
 
 // Success create a successful response
@@ -124,16 +83,16 @@ func Error(status int, message string, err error) *NormalResponse {
 	return resp
 }
 
-// Respond create a response
+// Respond creates a response.
 func Respond(status int, body interface{}) *NormalResponse {
 	var b []byte
-	var err error
 	switch t := body.(type) {
 	case []byte:
 		b = t
 	case string:
 		b = []byte(t)
 	default:
+		var err error
 		if b, err = json.Marshal(body); err != nil {
 			return Error(500, "body json marshal", err)
 		}
@@ -143,28 +102,6 @@ func Respond(status int, body interface{}) *NormalResponse {
 		status: status,
 		header: make(http.Header),
 	}
-}
-
-// RedirectResponse represents a redirect response.
-type RedirectResponse struct {
-	location string
-}
-
-// WriteTo writes to a response.
-func (r *RedirectResponse) WriteTo(ctx *models.ReqContext) {
-	ctx.Redirect(r.location)
-}
-
-// Status gets the response's status.
-// Required to implement api.Response.
-func (*RedirectResponse) Status() int {
-	return http.StatusFound
-}
-
-// Body gets the response's body.
-// Required to implement api.Response.
-func (r *RedirectResponse) Body() []byte {
-	return nil
 }
 
 func Redirect(location string) *RedirectResponse {
